pyproject.toml
[project]
name = "pyproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.0.27",
    "langchain-chroma>=0.2.0",
    "langchain-openai>=1.1.15",
    "python-dotenv>=1.0.1",
]
